test: test: migrate part of page-header tests
